% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis.R
\name{diff_analysis}
\alias{diff_analysis}
\title{Differential analysis}
\usage{
diff_analysis(
  exprset,
  project,
  is_count = TRUE,
  logFC_cut = 0,
  pvalue_cut = 1,
  adjpvalue_cut = 1,
  group = NULL,
  save = FALSE
)
}
\arguments{
\item{exprset}{expression matrix prepared by getmrnaexpr()/getmirnaexpr()/
getmrnaexpr_xena(), or your own expression matrix. If use your own
expression matrix, the argument "group" must be provided}

\item{project}{characters used as part of file name}

\item{is_count}{if the expression matrix is count. Default is TRUE. If TRUE,
the DEA will be done by DESeq2, egdeR and limma; if FALSE, the DEA will
be done by limma and wilcoxon test}

\item{logFC_cut}{threshold of logFC used for filtering the DEGs, default is 0}

\item{pvalue_cut}{p value cutpoint of the DEGs, default is 1}

\item{adjpvalue_cut}{adjusted p value cutpoint of the DEGs, default is 1}

\item{group}{a factor with two levels, specifying the group of samples,
which should be the same length as your sample size. If use your own
expression matrix, it must be provided}

\item{save}{save results to local, default is "FALSE"}
}
\value{
a list of differential analysis results
}
\description{
This function can automatically do differential analysis by 4
methods: DESeq2, egdeR, limma and wilcoxon test.
}
