% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_clinical.R
\name{dn_clinical}
\alias{dn_clinical}
\title{Download available clinical data of multiple / all cohorts.}
\usage{
dn_clinical(cohorts, filename = NULL)
}
\arguments{
\item{cohorts}{A character vector indicating the cohort(s) to query. See \code{\link{dn_cohorts}} for available cohorts.}
}
\value{
data.frame of patient clinical data of multiple / all cohorts.
}
\description{
Download available clinical data of multiple / all cohorts.
}
\examples{
cohorts = c("ACC", "BLCA", "COAD")
clinical = dn_clinical(cohorts)
}
\seealso{
\code{dn_clinical} uses the service \code{\link{dn_clinical_one}}. Note that for downloading all available barcodes, see \code{\link{dn_patient_barcodes}}, the object cohorts is a character vector containing all available cohort abbreviations.
}

