% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_clinical.R
\name{dn_gene_ID}
\alias{dn_gene_ID}
\title{Download all available genes.}
\usage{
dn_gene_ID(tcga_participant_barcode, page.Size, filename = NULL)
}
\arguments{
\item{tcga_participant_barcode}{A character containing a random TCGA patient barcode. See e.g. \code{\link{dn_patient_barcodes}} for available barcodes.}

\item{page.Size}{Number of records per page. Usually max is 2000.}
}
\value{
A character vector of all available gene symbols.
}
\description{
Download all available genes.
}
\seealso{
\code{dn_gene_all} uses the service \code{\link{Analyses.CopyNumber.Genes.All}}, see \code{\link{FirebrowseR}}.
}

