% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_miRNA.R
\name{dn_miRNA_ID}
\alias{dn_miRNA_ID}
\title{Download all miR names.}
\usage{
dn_miRNA_ID(tcga_participant_barcode, filename = NULL)
}
\arguments{
\item{tcga_participant_barcode}{A character containing a random TCGA patient barcode, the barcode "TCGA-BH-A1EO" can be used. See \code{\link{dn_patient_barcodes}} for available barcodes.}
}
\value{
A character vector containing all available miR names.
}
\description{
Download all miR names.
}
\seealso{
\code{dn_miRNA_ID} uses the service \code{\link{Samples.miRSeq}}, see \code{\link{FirebrowseR}}.
  It may happen that the data are NULL, the reasons can be:
    - There are no data for the specified barcode
    - For some cancer types (cohorts) there are no data at all
      -> try a tcga_participant_barcode of another cancer type
      Problem : without miRNA IDs there is no possibility to find out which cancers don´t have miRSeq expression values
      The good news: the output is the same for any adequate barcode
    - Lack of connection to the server -> try the same barcode again or choose another barcode
}

