% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_miRNA.R
\name{dn_miRSeq}
\alias{dn_miRSeq}
\title{Download sample-level log2 miRSeq expression values. Results may be filtered by mir, cohort and barcode. At least one miR must be supplied.}
\usage{
dn_miRSeq(mir, cohort, tcga_participant_barcode, sort_by, page.Size,
  filename = NULL)
}
\arguments{
\item{mir}{A character vector containing miR names, empty string queries all miR names. See \code{\link{miRNA_ID.R}} for available miR names.}

\item{cohort}{A character vector containing the cohort(s) to query, empty string queries all cohorts. See \code{\link{dn_cohorts}} for available cohorts.}

\item{tcga_participant_barcode}{A character vector containing TCGA barcodes, empty string queries all barcodes. See \code{\link{dn_patient_barcodes}} for available barcodes. Remark that the data are NULL for barcode(s) which isn´t (aren´t) barcode(s) of the specified cohort.}

\item{sort_by}{A character indicating the column which is used for sorting. The data can be sorted by tcga_participant_barcode, cohort, tool, mir and sample_type.}

\item{page.Size}{Number of records per page. Usually max is 2000.}
}
\value{
data.frame of log2 miRSeq expression values.
}
\description{
Download sample-level log2 miRSeq expression values. Results may be filtered by mir, cohort and barcode. At least one miR must be supplied.
}
\seealso{
\code{dn_miRSeq} uses the service \code{\link{Samples.miRSeq}}, see \code{\link{FirebrowseR}}.
}

