% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_miRNA.R
\name{dn_miRSeq_cohort}
\alias{dn_miRSeq_cohort}
\title{Download all available sample-level log2 miRSeq expression values of one cohort.}
\usage{
dn_miRSeq_cohort(cohort, page.Size, filename = NULL)
}
\arguments{
\item{cohort}{A character vector indicating the cohort to query. See \code{\link{dn_cohorts}} for available cohorts.}

\item{page.Size}{Number of records per page. Usually max is 2000.}
}
\value{
data.frame of sample-level log2 miRSeq expression values of the specified cohort.
}
\description{
Download all available sample-level log2 miRSeq expression values of one cohort.
}
\seealso{
\code{dn_miRSeq_cohort} uses the service \code{\link{dn_miRSeq}}.
}

