% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_mRNA.R
\name{dn_mRNA_ID}
\alias{dn_mRNA_ID}
\title{Download all mRNA IDs.}
\usage{
dn_mRNA_ID(tcga_participant_barcode, page.Size, filename = NULL)
}
\arguments{
\item{tcga_participant_barcode}{A character containing a random TCGA patient barcode, e.g. the barcode "TCGA-E9-A2JT" can be used.
It may happen that the data are NULL, the reasons can be:
    - There are no data for the specified barcode
    - For some cancer types (cohorts) there are no data at all -> try a tcga_participant_barcode of another cancer type
      Problem : without mRNA IDs there is no possibility to find out which cancers don´t have mRNARSeq expression values
      The good news: the output is the same for any adequate barcode
    - Lack of connection to the server -> try the same barcode again or choose another barcode
    -> please check if mRNA_ID, the output of \code{dn_mRNA_ID}, is NULL for the specified barcode before using the service
       \code{\link{dn_mRNASeq_cohort}}, otherwise the output of \code{\link{dn_mRNASeq_cohort}} is NULL as well}

\item{page.Size}{Number of records per page. Usually max is 2000.}
}
\value{
A character vector containing all available mRNA IDs.
}
\description{
Download all mRNA IDs.
}
\seealso{
\code{dn_mRNA_ID} uses the service \code{\link{Samples.mRNASeq}}, see \code{\link{FirebrowseR}}.
}

