% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_mRNA.R
\name{dn_mRNASeq_cohort}
\alias{dn_mRNASeq_cohort}
\title{Download all available sample-level log2 mRNASeq expression values of one cohort.}
\usage{
dn_mRNASeq_cohort(cohort, page.Size, filename = NULL)
}
\arguments{
\item{cohort}{A character vector containing the cohort to query. See \code{\link{dn_cohorts}} for available cohorts.}

\item{page.Size}{Number of records per page. Usually max is 2000. page.Size should be chosen bigger than the number of patients of the cohort.}
}
\value{
data.frame of sample-level log2 mRNASeq expression values of the specified cohort.
}
\description{
Download all available sample-level log2 mRNASeq expression values of one cohort.
}
\examples{
cohort = "ESCA"
page_size = 2000
esca.mRNASeq = dn_mRNASeq_cohort(cohort, page.Size)
}
\seealso{
\code{dn_mRNASeq_cohort} uses the service \code{\link{dn_mRNASeq}}.
}

