% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_mutation_SMG.R
\name{dn_mutation.SMG.Exp}
\alias{dn_mutation.SMG.Exp}
\title{Download significantly mutated genes, as scored by MutSig. Results may be filtered by cohort, rank, gene, tool and/or Q-value threshold. At least one cohort or gene must be supplied.}
\usage{
dn_mutation.SMG.Exp(gene, rank, cohort, page.Size, sort_by, filename = NULL)
}
\arguments{
\item{gene}{A character vector of gene symbols.}

\item{rank}{Number of significant genes to return.}

\item{cohort}{A character vector indicating the cohort(s) to query, empty string queries all cohorts. See \code{\link{dn_cohorts}} for available cohorts.}

\item{page.Size}{Number of records per page.  Usually max is 2000.}

\item{sort_by}{A character indicating the column which is used for sorting. Data can be sorted by cohort, tool, gene and rank.}
}
\value{
data.frame of selected columns from the MAF generated by MutSig.
}
\description{
Download significantly mutated genes, as scored by MutSig. Results may be filtered by cohort, rank, gene, tool and/or Q-value threshold. At least one cohort or gene must be supplied.
}
\examples{
gene = c("A1BG", "A1CF", "A2M")
rank = ""
cohort = "ACC"
page.Size = 25
sort_by = "gene"
obj = dn_mutation.SMG.Exp(gene, rank, cohort, page.Size, sort_by)
}
\seealso{
\code{dn_mutation_SMG.Exp} uses the service \code{\link{Analyses.Mutation.SMG}}, see \code{\link{FirebrowseR}}.
}

