% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_mutation_SMG.R
\name{dn_mutation.SMG_cohort}
\alias{dn_mutation.SMG_cohort}
\title{Download selected columns from the MAF generated by MutSig for one cohort.}
\usage{
dn_mutation.SMG_cohort(cohort, page.Size, filename = NULL)
}
\arguments{
\item{cohort}{A character vector indicating the cohort to query. See \code{\link{dn_cohorts}} for available cohorts.}

\item{page.Size}{Number of records per page.  For cancers with small number of patients use a small page.Size, e.g. page.Size = 25. If there is an error "Error in mut.cohort[[page.Counter]]: subscript out of bounds", try a smaller page.Size.}
}
\value{
data.frame of selected columns from the MAF generated by MutSigof for all patient barcodes of the specified cohort.
}
\description{
Download selected columns from the MAF generated by MutSig for one cohort.
}
\seealso{
\code{dn_mutation_cohort} uses the service \code{\link{dn_mutation.Exp}}.
}

