% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_mutation.R
\name{dn_mutation_cohort}
\alias{dn_mutation_cohort}
\title{Download selected columns from the MAF generated by MutSig for one cohort.}
\usage{
dn_mutation_cohort(cohort, page.Size, filename = NULL)
}
\arguments{
\item{cohort}{A character vector containing the cohort to query. See \code{\link{dn_cohorts}} for available cohorts.}

\item{page.Size}{Number of records per page. Usually max is 2000. For cancers with small number of patients use a small page.Size, e.g. page.Size = 250.}
}
\description{
Download selected columns from the MAF generated by MutSig for one cohort.
}
\seealso{
\code{dn_mutation_cohort} uses the service \code{\link{dn_mutation.Exp}}.
}

