% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_mutation_SMG.R
\name{dn_mutation_SMG.Exp}
\alias{dn_mutation_SMG.Exp}
\title{Download significantly mutated genes, as scored by MutSig. It may be filtered by cohort, rank, gene, tool and/or Q-value threshold. At least one cohort or gene must be supplied.}
\usage{
dn_mutation_SMG.Exp(gene, rank, cohort, page.Size, sort_by, filename = NULL)
}
\arguments{
\item{gene}{A character vector of gene symbols. See \code{\link{dn_gene_ID}} for available gene symbols.}

\item{rank}{Number of significant genes to return.}

\item{cohort}{A character vector indicating the cohort(s) to query. See \code{\link{dn_cohorts}} for available cohorts.}

\item{page.Size}{Number of records per page.}

\item{sort_by}{A character indicating the to column which is used for sorting. Data can be sorted by q, cohort, tool, gene and rank.}
}
\value{
data.frame of selected columns from the MAF generated by MutSig.
}
\description{
Download significantly mutated genes, as scored by MutSig. It may be filtered by cohort, rank, gene, tool and/or Q-value threshold. At least one cohort or gene must be supplied.
}
\seealso{
\code{dn_mutation_SMG.Exp} uses the service \code{\link{Analyses.Mutation.SMG}}, see \code{\link{FirebrowseR}}.
}

