% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_clinical.R
\name{dn_patient_barcodes}
\alias{dn_patient_barcodes}
\title{Download patient barcodes with corresponding cohort abbreviations of cohort(s) to query.}
\usage{
dn_patient_barcodes(cohort, filename = NULL)
}
\arguments{
\item{cohort}{A character vector indicating the cohort(s) to query. See \code{\link{dn_cohorts}} for available cohorts.}
}
\value{
A data.frame containing TCGA barcodes of the cohort(s) with corresponding cohort abbreviations.
}
\description{
Download patient barcodes with corresponding cohort abbreviations of cohort(s) to query.
}
\seealso{
\code{\link{dn_clinical_all}}
}

