% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcnv.R
\name{getmethybeta}
\alias{getmethybeta}
\title{Get GDC TCGA 450K DNA methylation beta value matirx}
\usage{
getmethybeta(project)
}
\arguments{
\item{project}{valid TCGA project name(s) from 33 TCGA projects
\itemize{
\item{ TCGA-ACC }
\item{ TCGA-BLCA }
\item{ TCGA-BRCA }
\item{ TCGA-CESC }
\item{ TCGA-CHOL }
\item{ TCGA-COAD }
\item{ TCGA-DLBC }
\item{ TCGA-ESCA }
\item{ TCGA-GBM }
\item{ TCGA-HNSC }
\item{ TCGA-KICH }
\item{ TCGA-KIRC }
\item{ TCGA-KIRP }
\item{ TCGA-LAML }
\item{ TCGA-LGG }
\item{ TCGA-LIHC }
\item{ TCGA-LUAD }
\item{ TCGA-LUSC }
\item{ TCGA-MESO }
\item{ TCGA-OV }
\item{ TCGA-PAAD }
\item{ TCGA-PCPG }
\item{ TCGA-PRAD }
\item{ TCGA-READ }
\item{ TCGA-SARC }
\item{ TCGA-SKCM }
\item{ TCGA-STAD }
\item{ TCGA-TGCT }
\item{ TCGA-THCA }
\item{ TCGA-THYM }
\item{ TCGA-UCEC }
\item{ TCGA-UCS }
\item{ TCGA-UVM }
}}
}
\value{
450k DNA methylation beta value matrix and clinical information and
probe information. The data are saved in the directory of "output_methy".
}
\description{
This function can easily get the 450K DNA methylation beta value
matrix and the corresponding clinical information from GDC TCGA.
}
