% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmrnaexpr_xena.R
\name{getmrnaexpr_xena}
\alias{getmrnaexpr_xena}
\title{Get mRNA and lncRNA expression matrix from XENA file}
\usage{
getmrnaexpr_xena(file_name)
}
\arguments{
\item{file_name}{gene expression RNAseq(HTSeq-Counts) file downloaded from
https://gdc.xenahubs.net}
}
\value{
mRNA and lncRNA expression matrix. The data are saved in the
directory of "output_mRNA_expr_xena".
}
\description{
Get mRNA and lncRNA expression matrix from XENA file
}
