% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpancancer_xena.R
\name{getpancancer_xena}
\alias{getpancancer_xena}
\title{Prepare pan-cancer expression matrix and clinical information from XENA}
\usage{
getpancancer_xena(
  tcga_expr_file,
  tcga_clin_file,
  gtex_expr_file,
  gtex_pheno_file,
  type = c("tcga", "gtex", "tcga+gtex"),
  save_lnc = TRUE
)
}
\arguments{
\item{tcga_expr_file}{TCGA pan-cancer gene expression RNA-seq file}

\item{tcga_clin_file}{TCGA clin file}

\item{gtex_expr_file}{GTEx gene expression RNA-seq file}

\item{gtex_pheno_file}{GTEx pheno file}

\item{type}{type of pan-cancer. "tcga" for only TCGA pan-cancer,
tcga_expr_file and tcga_clin_file should be provided. "gtex" for only GTEx
pan-cancer, gtex_expr_file and gtex_pheno_file should be provided.
"tcga+gtex" for combined TCGA and GTEX pan-cancer, all the 4 files should
be provided.}

\item{save_lnc}{save lncRNA data, default is "TRUE"}
}
\value{
pan-cancer expression matrix and clinical info. The data are saved
under the directory of "output_pancancer_xena".
}
\description{
This function can help you prepare the pan-cancer expression
matrix and clinical(phenotype) information.
}
\details{
All the files must be downloaded from the following urls:
\itemize{
\item{tcga_expr_file: https://toil-xena-hub.s3.us-east-1.amazonaws.com/
download/tcga_RSEM_gene_tpm.gz}
\item{gtex_expr_file: https://toil-xena-hub.s3.us-east-1.amazonaws.com/
download/gtex_RSEM_gene_tpm.gz}
\item{tcga_clin_file: https://tcga-pancan-atlas-hub.s3.us-east-1.
amazonaws.com/download/Survival_SupplementalTable_S1_20171025_xena_sp}
\item{gtex_pheno_file: https://toil-xena-hub.s3.us-east-1.amazonaws.com/
download/GTEX_phenotype.gz}
}
}
