% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsnvmaf.R
\name{getsnvmaf}
\alias{getsnvmaf}
\title{Get GDC TCGA MAF files}
\usage{
getsnvmaf(project)
}
\arguments{
\item{project}{one of 33 TCGA projects
\itemize{
\item{ TCGA-ACC }
\item{ TCGA-BLCA }
\item{ TCGA-BRCA }
\item{ TCGA-CESC }
\item{ TCGA-CHOL }
\item{ TCGA-COAD }
\item{ TCGA-DLBC }
\item{ TCGA-ESCA }
\item{ TCGA-GBM }
\item{ TCGA-HNSC }
\item{ TCGA-KICH }
\item{ TCGA-KIRC }
\item{ TCGA-KIRP }
\item{ TCGA-LAML }
\item{ TCGA-LGG }
\item{ TCGA-LIHC }
\item{ TCGA-LUAD }
\item{ TCGA-LUSC }
\item{ TCGA-MESO }
\item{ TCGA-OV }
\item{ TCGA-PAAD }
\item{ TCGA-PCPG }
\item{ TCGA-PRAD }
\item{ TCGA-READ }
\item{ TCGA-SARC }
\item{ TCGA-SKCM }
\item{ TCGA-STAD }
\item{ TCGA-TGCT }
\item{ TCGA-THCA }
\item{ TCGA-THYM }
\item{ TCGA-UCEC }
\item{ TCGA-UCS }
\item{ TCGA-UVM }
}}
}
\value{
maf file and clinical information, which can be directly used by
maftools::read.maf(). The data are saved in the directory of "output_snv"
}
\description{
This function can automatically download and prepare the TCGA
maf files(masked somatic mutation) with the corresponding clinical
information. The output can be directly used by maftools::read.maf()
function.
}
