% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_clinical.R
\name{patient_barcodes}
\alias{patient_barcodes}
\title{Extract TCGA patient barcodes from clinical data.}
\usage{
patient_barcodes(clinical)
}
\arguments{
\item{clinical}{A data frame containing clinical data. See \code{\link{dn_clinical}}.}
}
\value{
A character vector containing TCGA barcodes.
}
\description{
Extract TCGA patient barcodes from clinical data.
}
\examples{
cohort = "BRCA"
brca.clinical = dn_clinical_one(cohort)
brca.barcodes = patient_barcodes(clinical) # returns all patient barcodes of the cohort BRCA
}

