% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene_paired}
\alias{plot_gene_paired}
\title{Box plot of marker expression between paired normal and tumor samples}
\usage{
plot_gene_paired(exprset, marker)
}
\arguments{
\item{exprset}{expression matrix}

\item{marker}{marker(s) you want to display}
}
\value{
box plot and the plotting data
}
\description{
Box plot of marker expression between paired normal and tumor samples
}
