% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape.mRNASeq}
\alias{reshape.mRNASeq}
\title{Reshape sample-level log2 mRNASeq expression values of the sample type TP. For reshaping all sample-level log2 
mRNASeq expression values of one cohort, see \code{\link{dn_mRNASeq_cohort}}, use \code{\link{reshape.mRNASeq_geneID}}.}
\usage{
reshape.mRNASeq(data, sample_type = "TP")
}
\arguments{
\item{data}{data.frame of log2 mRNASeq expression values. See \code{\link{dn_mRNASeq}} for downloading these values.}
}
\value{
nxp-matrix, n = number of patients, p = number of genes, \cr
        (matrix(i,j))_{i,j} = (matrix(i,j))_{i,j} =  (expression_log2)_{ij}, where \cr
        data$tcga_participant_barcode==barcode[i], data$gene==gene[j], \cr
        i=1,..,n, j=1,...,p
}
\description{
Reshape sample-level log2 mRNASeq expression values of the sample type TP. For reshaping all sample-level log2 
mRNASeq expression values of one cohort, see \code{\link{dn_mRNASeq_cohort}}, use \code{\link{reshape.mRNASeq_geneID}}.
}
\examples{
gene = c("A1BG", "AAAS")
cohort = "ESCA"
tcga_participant_barcode = c("TCGA-2H-A9GF", "TCGA-LN-A49M") #  TCGA patient barcodes from ESCA
sort_by = "gene"
page.Size = 250
obj = dn_mRNASeq(gene, cohort, tcga_participant_barcode, sort_by, page.Size)
obj_reshaped = reshape.mRNASeq(obj, sample_type = "TP")
}

