% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_cdr_surv.R
\docType{data}
\name{tcga_cdr_surv}
\alias{tcga_cdr_surv}
\title{Curated clinical data}
\format{
An object of class \code{data.frame} with 12591 rows and 34 columns.
}
\usage{
tcga_cdr_surv
}
\description{
Curated clinical data from the Pan-cancer Atlas paper titled "An Integrated
TCGA Pan-Cancer Clinical Data Resource (TCGA-CDR) to drive high quality survival
outcome analytics". The paper highlights four types of carefully curated survival
endpoints, and recommends the use of the endpoints of OS, PFI, DFI, and DSS
for each TCGA cancer type.
}
\details{
\itemize{
\item OS: overall survial
\item PFI: progression-free interval
\item DSS: disease-specific survival
\item DFI: disease-free interval
}

You can download it from \href{https://tcga-pancan-atlas-hub.s3.us-east-1.amazonaws.com/download/Survival_SupplementalTable_S1_20171025_xena_sp}{XENA}
}
\keyword{datasets}
