% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_survival.R
\name{batch_survival}
\alias{batch_survival}
\title{Batch survival analysis}
\usage{
batch_survival(
  exprset,
  clin,
  expr_type = c("counts", "tpm", "fpkm"),
  project,
  min_sample_size = 5,
  print_index = TRUE
)
}
\arguments{
\item{exprset}{expression set prepared by getmrnaexpr()}

\item{clin}{clinical information prepared by getmrnaexpr()}

\item{expr_type}{type of expression set}

\item{project}{a TCGA project}

\item{min_sample_size}{min sample size}

\item{print_index}{print index, default is "TRUE"}
}
\value{
survival analysis resaults
}
\description{
Batch survival analysis
}
