% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_survival.R
\name{batch_survival}
\alias{batch_survival}
\title{Batch survival analysis}
\usage{
batch_survival(
  exprset,
  clin,
  expr_type = c("counts", "tpm", "fpkm"),
  optimal_cut = TRUE,
  project,
  min_sample_size = 5,
  print_index = TRUE
)
}
\arguments{
\item{exprset}{Expression matrix from getmrnaexpr() or getmirnaexpr().}

\item{clin}{Clinical information from getmrnaexpr() or getmirnaexpr().}

\item{expr_type}{Type of expression matrix, one of count, tpm, fpkm.
If "count", the expression matrix will be transformed by DESeq2::vst() or
DESeq2::varianceStabilizingTransformation. For other types, the expression
will be transformed by log2(x + 0.1).}

\item{optimal_cut}{Use "optimal" cutpoint to do survival analysis. If FALSE,
median of gene expression will be used. Optimal cutpoint is calculated by
survminer::surv_cutpoint().}

\item{project}{Characters used as the part of file name when saving results.}

\item{min_sample_size}{Min sample size of each group for survival analysis,
default is 5.}

\item{print_index}{Print index, default is TRUE.}
}
\value{
a list of survival analysis results, both log-rank and cox regression
}
\description{
This function can automatically do log-rank test and univariate
cox regression according to the "optimal" cutpoint or median of gene
expression. Support mRNA, lncRNA and miRNA expression matrix(count, tpm,
fpkm are all supported).
}
