% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis.R
\name{diff_analysis}
\alias{diff_analysis}
\title{Differential analysis}
\usage{
diff_analysis(exprset, project, save = FALSE)
}
\arguments{
\item{exprset}{expression set prepared by getmrnaexpr()}

\item{project}{a TCGA project}

\item{save}{save results to local, defautl is "FALSE"}
}
\value{
the differential analysis results
}
\description{
Differential analysis
}
