% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis.R
\name{diff_analysis}
\alias{diff_analysis}
\title{Differential analysis}
\usage{
diff_analysis(exprset, project, save = FALSE)
}
\arguments{
\item{exprset}{counts expression matrix prepared by getmrnaexpr()}

\item{project}{a TCGA project}

\item{save}{save results to local, defautl is "FALSE"}
}
\value{
a list of differential analysis results
}
\description{
This function can automatically do differential analysis by 3
ways: DESeq2, egdeR and limma.
}
