% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_miRNA.R
\name{dn_miRSeq}
\alias{dn_miRSeq}
\title{Download sample-level log2 miRSeq expression values. Results may be filtered by mir, cohort and barcode.}
\usage{
dn_miRSeq(mir, cohort, tcga_participant_barcode, page.Size, sort_by,
  filename = NULL)
}
\arguments{
\item{mir}{A character vector containing miR names, empty string or mir=miRNA_ID queries all mirs. At least one mir must be supplied. See \code{\link{miRNA_ID.R}} for available miR names.}

\item{cohort}{A character vector containing the cohort(s) to query, empty string queries all cohorts. See \code{\link{dn_cohorts}} for available cohorts.}

\item{tcga_participant_barcode}{A character vector containing TCGA barcodes, empty string queries all barcodes. See \code{\link{patient_barcodes}} for available barcodes. Remark that the data can be NULL for barcode(s) which isn´t (aren´t) barcode(s) of the specified cohort.}

\item{page.Size}{Number of records per page. Usually max is 2000.}

\item{sort_by}{A character indicating the column which is used for sorting. The data can be sorted by tcga_participant_barcode, cohort, tool, mir and sample_type.}
}
\value{
data.frame of log2 miRSeq expression values.
}
\description{
Download sample-level log2 miRSeq expression values. Results may be filtered by mir, cohort and barcode.
}
\examples{
mir = miRNA_ID[1:10]
cohort = "BLCA"
tcga_participant_barcode = "TCGA-ZF-AA53"  # TCGA patient barcode from BLCA
page.Size = 250
sort_by =  "tcga_participant_barcode"
obj = dn_miRSeq(mir, cohort, tcga_participant_barcode, page.Size, sort_by)
}
\seealso{
\code{dn_miRSeq} uses the service \code{\link{Samples.miRSeq}}, see \code{\link{FirebrowseR}}.
}

