% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_mRNA.R
\name{dn_mRNASeq}
\alias{dn_mRNASeq}
\title{Download sample-level log2 mRNASeq expression values. Results may be filtered by gene, tcga_participant_barcode and cohort.}
\usage{
dn_mRNASeq(gene, cohort, tcga_participant_barcode, page.Size, sort_by,
  filename = NULL)
}
\arguments{
\item{gene}{A character vector of gene symbols, empty string or gene = mRNA_ID queries all genes. At least one gene must be supplied. See \code{\link{mRNA_ID.R}} for available genes.}

\item{cohort}{A character vector indicating cohort(s) to query, empty string queries all cohorts. See \code{\link{dn_cohorts}} for available cohorts.}

\item{tcga_participant_barcode}{A character vector containing TCGA barcodes, empty string queries all barcodes. See e.g. \code{\link{patient_barcodes}} for available barcodes. Remark that the data can be NULL for barcode(s) which isn´t (aren´t) barcode(s) of the specified cohort.}

\item{page.Size}{Number of records per page. Usually max is 2000.}

\item{sort_by}{A character indicating the column which is used for sorting. The data can be sorted by tcga_participant_barcode, cohort, gene, protocol and sample_type.}
}
\value{
data.frame of log2 mRNASeq expression values.
}
\description{
Download sample-level log2 mRNASeq expression values. Results may be filtered by gene, tcga_participant_barcode and cohort.
}
\examples{
gene = "AAAS" 
cohort = "ESCA"
tcga_participant_barcode = c("TCGA-2H-A9GF", "TCGA-LN-A49M") #  TCGA patient barcodes from ESCA
sort_by = "gene"
page.Size = 250
obj = dn_mRNASeq(gene, cohort, tcga_participant_barcode, sort_by, page.Size2)
}
\seealso{
\code{dn_mRNASeq} uses the service \code{\link{Samples.mRNASeq}}, see \code{\link{FirebrowseR}}.
}

