% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_mRNA.R
\name{dn_mRNASeq_cohort}
\alias{dn_mRNASeq_cohort}
\title{Download all available sample-level log2 mRNASeq expression values of one cohort.}
\usage{
dn_mRNASeq_cohort(cohort, page.Size, filename = NULL)
}
\arguments{
\item{cohort}{A character vector containing the cohort to query. See \code{\link{dn_cohorts}} for available cohorts.}

\item{page.Size}{Number of records per page. Usually max is 2000. For cancers with small number of patients use a small page.Size, e.g. page.Size = 250.}

\item{mRNA_ID}{A character vector containing gene symbols. See \code{\link{mRNA_ID.R}} for available gene symbols.}
}
\value{
data.frame of sample-level log2 mRNASeq expression values of the specified cohort.
}
\description{
Download all available sample-level log2 mRNASeq expression values of one cohort.
}
\seealso{
\code{dn_mRNASeq_cohort} uses the service \code{\link{dn_mRNASeq}}.
}

