% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_mutation.R
\name{dn_mutation.Exp}
\alias{dn_mutation.Exp}
\title{Download selected columns from the MAF generated by MutSig. Results may be filtered by gene, cohort, barcode and tool. At least one gene or barcode or cohort must be supplied.}
\usage{
dn_mutation.Exp(tcga_participant_barcode, cohort, gene, page.Size, sort_by,
  filename = NULL)
}
\arguments{
\item{tcga_participant_barcode}{A character vector containing TCGA barcodes, empty string queries all barcodes. See \code{\link{patient_barcodes}} for available barcodes. Remark that the data are NULL for barcode(s) which isn´t (aren´t) barcode(s) of the specified cohort.}

\item{cohort}{A character vector containing the cohort(s) to query. See \code{\link{dn_cohorts}} for available cohorts.}

\item{gene}{A character vector containing gene symbols.}

\item{page.Size}{Number of records per page. For downloading data of single barcodes or genes the page.Size should be chosen small, i.e. 25, but usually not larger than 1000.}

\item{sort_by}{A character indicating the column which is used for sorting. The data can be sorted by tcga_participant_barcode, cohort, tool and gene.}
}
\value{
data.frame of selected columns from the MAF generated by MutSig.
}
\description{
Download selected columns from the MAF generated by MutSig. Results may be filtered by gene, cohort, barcode and tool. At least one gene or barcode or cohort must be supplied.
}
\examples{
tcga_participant_barcode = "TCGA-AG-A002" # a READ patient barcode
cohort = "READ"
gene = c("A1BG", "A1CF", "A2M")
page.Size = 250
sort_by = "gene"
obj = dn_mutation.Exp(tcga_participant_barcode, cohort, gene, page.Size, sort_by)
}
\seealso{
\code{dn_mutation.Exp} uses the service \code{\link{Analyses.Mutation.MAF}}, see \code{\link{FirebrowseR}}.
}

