% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_mutation.R
\name{dn_mutation_cohort}
\alias{dn_mutation_cohort}
\title{Download selected columns from the MAF generated by MutSig produced by the selected Firehose tool for one cohort.}
\usage{
dn_mutation_cohort(cohort, page.Size, tool, filename = NULL)
}
\arguments{
\item{cohort}{A character vector containing the cohort to query. See \code{\link{dn_cohorts}} for available cohorts.}

\item{page.Size}{Number of records per page. Usually max is 2000. For cancers with small number of patients use rather a small page.Size, e.g. page.Size = 250.}

\item{tool}{Narrow search to include only data/results produced by the selected Firehose tool. Multiple values are allowed MutSig2.0,MutSig2CV. Default value is MutSig2CV.}
}
\description{
Download selected columns from the MAF generated by MutSig produced by the selected Firehose tool for one cohort.
}
\examples{
cohort = "READ"
page.Size = 1000
tool = "MutSig2CV"
read.mutation = dn_mutation_cohort(cohort, page.Size, tool)
}
\seealso{
\code{dn_mutation_cohort} uses the service \code{\link{dn_mutation.Exp}}.
}

