% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcnv.R
\name{getmethybeta}
\alias{getmethybeta}
\title{Get GDC TCGA 450K DNA methylation beta value matirx}
\usage{
getmethybeta(project)
}
\arguments{
\item{project}{valid TCGA project name(s) from 33 TCGA projects:
\itemize{
\item TCGA-ACC
\item TCGA-BLCA
\item TCGA-BRCA
\item TCGA-CESC
\item TCGA-CHOL
\item TCGA-COAD
\item TCGA-DLBC
\item TCGA-ESCA
\item TCGA-GBM
\item TCGA-HNSC
\item TCGA-KICH
\item TCGA-KIRC
\item TCGA-KIRP
\item TCGA-LAML
\item TCGA-LGG
\item TCGA-LIHC
\item TCGA-LUAD
\item TCGA-LUSC
\item TCGA-MESO
\item TCGA-OV
\item TCGA-PAAD
\item TCGA-PCPG
\item TCGA-PRAD
\item TCGA-READ
\item TCGA-SARC
\item TCGA-SKCM
\item TCGA-STAD
\item TCGA-TGCT
\item TCGA-THCA
\item TCGA-THYM
\item TCGA-UCEC
\item TCGA-UCS
\item TCGA-UVM
}

If you provide more than one TCGA project names, it will combine the data.
This is useful when you try to combine different cancer types, such as
TCGA-COAD and TCGA-READ.}
}
\value{
rdata files which are saved in the directory of "output_methy".
\itemize{
\item \strong{TCGA-XXX_methy_beta_SummarizedExperiment.rdata}: SummarizedExperiment
object, all the other files are extracted from this object.
\item \strong{TCGA-XXX_clinicalSE.rdata}: indexed clinical data extracted from the
SummarizedExperiment object.
\item \strong{TCGA-XXX_probe_info.rdata}: probe information
\item \strong{TCGA-XXX_methy_beta_expr.rdata}: DNA methylation beta value matrix
\item \strong{TCGA-XXX_pd.rdata}: a simple pd file including sample name and sample
type(normal or tumor), which can be used in \code{ChAMP}.
}
}
\description{
This function can automatically query, download and prepare the
450K DNA methylation beta value matrix and the corresponding clinical
information from GDC TCGA.
}
\details{
In GDC database the clinical data can be retrieved from different sources:
\itemize{
\item indexed clinical: a refined clinical data that is created using the XML files
\item XML files: original source of the data
\item BCR Biotab: tsv files parsed from XML files
}

the clinical data extracted from the SummarizedExperiment object is indexed
clinical, and subtype information from marker TCGA papers are added, these
information have prefix "paper_". more information about clinical data,
please see: \href{https://bioconductor.org/packages/release/bioc/vignettes/TCGAbiolinks/inst/doc/clinical.html}{clinical section} of the \code{TCGAbiolinks} vignette.
}
\references{
Colaprico A, Silva TC, Olsen C, Garofano L, Cava C, Garolini D,
Sabedot T, Malta TM, Pagnotta SM, Castiglioni I, Ceccarelli M, Bontempi G,
Noushmehr H (2015). “ TCGAbiolinks: An R/Bioconductor package for integrative
analysis of TCGA data.” Nucleic Acids Research.
}
