% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmirnaexpr.R
\name{getmirnaexpr}
\alias{getmirnaexpr}
\title{Get GDC TCGA miRNA expression set}
\usage{
getmirnaexpr(project)
}
\arguments{
\item{project}{a TCGA project}
}
\value{
expression set
}
\description{
Get GDC TCGA miRNA expression set
}
