% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmrnaexpr.R
\name{getmrnaexpr}
\alias{getmrnaexpr}
\title{Get GDC TCGA mRNA and lncRNA expression set and clinical info}
\usage{
getmrnaexpr(project)
}
\arguments{
\item{project}{a TCGA project}
}
\value{
expression set and clinical info
}
\description{
Get GDC TCGA mRNA and lncRNA expression set and clinical info
}
