% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmrnaexpr_xena.R
\name{getmrnaexpr_xena}
\alias{getmrnaexpr_xena}
\title{Get mRNA and lncRNA expression matrix from XENA file}
\usage{
getmrnaexpr_xena(expr_file, clin_file = NULL)
}
\arguments{
\item{expr_file}{gene expression RNAseq file downloaded from XENA(gdchub)}

\item{clin_file}{phenotype file downloaded from XENA(gdchub)}
}
\value{
mRNA, lncRNA expression matrix and clinical info. The data are saved
under the directory of "output_mRNA_expr_xena".
}
\description{
Get mRNA and lncRNA expression matrix from XENA file
}
