% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Box plot of marker expression between groups}
\usage{
plot_gene(exprset, marker, group)
}
\arguments{
\item{exprset}{expression matrix}

\item{marker}{marker(s) you want to display}

\item{group}{the groups of your samples}
}
\value{
box plot and the plotting data
}
\description{
Box plot of marker expression between groups
}
