% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape.mRNASeq}
\alias{reshape.mRNASeq}
\title{Reshape sample-level log2 mRNASeq expression values of the sample type TP. Primarily this service is intended to reshape all sample-level log2 mRNASeq expression values of one cohort.
See \code{\link{dn_mRNASeq_cohort}}.}
\usage{
reshape.mRNASeq(data, sample_type = "TP")
}
\arguments{
\item{data}{data.frame of log2 mRNASeq expression values. See \code{\link{dn_mRNASeq}} for downloading these values.}
}
\value{
nxp-matrix, n = number of patients, p = number of genes, (matrix(i,j))_{i,j} = ((data$tcga_participant_barcode==barcode[i], data$gene==gene[j]))_{i,j}
}
\description{
Reshape sample-level log2 mRNASeq expression values of the sample type TP. Primarily this service is intended to reshape all sample-level log2 mRNASeq expression values of one cohort.
See \code{\link{dn_mRNASeq_cohort}}.
}
\examples{
cohort = "ESCA"
page_size = 2000
esca.mRNASeq = dn_mRNASeq_cohort(cohort, page.Size)
esca.mRNASeq_reshaped = reshape.mRNASeq (esca.mRNASeq, sample_type = "TP")
}

